Prepares a Correlation Table

Reads a data frame and presents Pearson correlations above the diagonal and Spearman correlations below.

prepare_correlation_table(df, digits = 2, bold = 0.05, format = "html",
  ...)

Arguments

df
Data frame containing at least two variables that are either numeric or logical and at least five observations.

digits
The number of digits that you want to report.

bold
Indicate the p-Value for for identifying significant correlations in bold brint. Defaults to 0.05. If set to 0, no bold print is being used.

format
The format that you want knitr::kable to produce ("html" or "latex")

...
Additional parameters that are passed on to knitr::kable

Value

A list containing four items:

"df_corr"

A data frame containg the correlations

"df_prob"

A data frame containg the p-values of the correlations

"df_n"

A data frame containg the number of observations used for the correlations

"kable_ret"

The return value provided by kable() containing the formatted table

Examples

t <- prepare_correlation_table(mtcars)
t$kable_ret

#> <table>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:right;"> A </th>
#>    <th style="text-align:right;"> B </th>
#>    <th style="text-align:right;"> C </th>
#>    <th style="text-align:right;"> D </th>
#>    <th style="text-align:right;"> E </th>
#>    <th style="text-align:right;"> F </th>
#>    <th style="text-align:right;"> G </th>
#>    <th style="text-align:right;"> H </th>
#>    <th style="text-align:right;"> I </th>
#>    <th style="text-align:right;"> J </th>
#>    <th style="text-align:right;"> K </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> A: mpg </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.85</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.85</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.78</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.68</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.87</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.42</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.66</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.60</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.48</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.55</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> B: cyl </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.91</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.90</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.83</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.70</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.78</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.59</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.81</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.52</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.49</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.53</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> C: disp </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.91</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.93</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.79</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.71</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.89</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.43</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.71</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.59</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.56</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.39</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> D: hp </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.89</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.90</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.85</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.45</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.66</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.71</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.72</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.24</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.13</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.75</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> E: drat </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.65</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.68</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.68</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.52</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.71</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.09</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.44</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.71</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.70</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.09</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> F: wt </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.89</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.86</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.90</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.77</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.75</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style="     ">-0.17</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.55</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.69</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.58</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.43</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> G: qsec </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.47</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.57</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.46</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.67</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.09</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.23</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.74</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.23</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.21</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.66</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> H: vs </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.71</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.81</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.72</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.75</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.45</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.59</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.79</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style="     ">0.17</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.21</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.57</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> I: am </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.56</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.52</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.62</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.36</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.69</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.74</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.20</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.17</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.79</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.06</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> J: gear </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.54</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.56</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.59</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.33</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.74</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.68</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.15</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.28</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.81</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> <span style="     ">0.27</span> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> K: carb </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.66</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.58</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.54</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.73</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.13</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">0.50</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.66</span> </td>
#>    <td style="text-align:right;"> <span style=" font-weight: bold;    ">-0.63</span> </td>
#>    <td style="text-align:right;"> <span style="     ">-0.06</span> </td>
#>    <td style="text-align:right;"> <span style="     ">0.11</span> </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#> </tbody>
#> <tfoot><tr><td style="padding: 0; border: 0;" colspan="100%">
#> <sup></sup> This table reports Pearson correlations above and Spearman correlations below the diagonal. Number of observations: 32. Correlations with significance levels below 5% appear in bold print.</td></tr></tfoot>
#> </table>