Prepares a Table of Descriptive Statistics

Reads a data frame and reports descriptive statistics (n, mean, standard deviation, minimum, first quartile, median, third quartile, maximum) for all members of the data frame that are eiter numeric or logical.

prepare_descriptive_table(df, digits = c(0, 3, 3, 3, 3, 3, 3, 3),
  format = "html")

Arguments

df
Data frame containing at least one variable that is either numeric or logical and at least two observations.

digits
Number of decimal digits that you want to be displayed for each column. If you provide NA, then the column is ommited from the output.

format
character scalar that is handed over to knitr::kabel (e.g., "html" or "latex").

Value

A list containing two items.

"df"

A data frame containg the descriptive table

"kable_ret"

The return value provided by kable() containing the formatted table

Details

The digits parameter from prepare_descriptive_table() uses the default method of kable() to format numbers, calling round(). This implies that trailing zeroes are just ommitted.

Examples

prepare_descriptive_table(mtcars)$kable_ret

#> <table>
#> <caption>Descriptive Statistics</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:right;"> N </th>
#>    <th style="text-align:right;"> Mean </th>
#>    <th style="text-align:right;"> Std. dev. </th>
#>    <th style="text-align:right;"> Min. </th>
#>    <th style="text-align:right;"> 25 % </th>
#>    <th style="text-align:right;"> Median </th>
#>    <th style="text-align:right;"> 75 % </th>
#>    <th style="text-align:right;"> Max. </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> mpg </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 20.091 </td>
#>    <td style="text-align:right;"> 6.027 </td>
#>    <td style="text-align:right;"> 10.400 </td>
#>    <td style="text-align:right;"> 15.425 </td>
#>    <td style="text-align:right;"> 19.200 </td>
#>    <td style="text-align:right;"> 22.80 </td>
#>    <td style="text-align:right;"> 33.900 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> cyl </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 6.188 </td>
#>    <td style="text-align:right;"> 1.786 </td>
#>    <td style="text-align:right;"> 4.000 </td>
#>    <td style="text-align:right;"> 4.000 </td>
#>    <td style="text-align:right;"> 6.000 </td>
#>    <td style="text-align:right;"> 8.00 </td>
#>    <td style="text-align:right;"> 8.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> disp </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 230.722 </td>
#>    <td style="text-align:right;"> 123.939 </td>
#>    <td style="text-align:right;"> 71.100 </td>
#>    <td style="text-align:right;"> 120.825 </td>
#>    <td style="text-align:right;"> 196.300 </td>
#>    <td style="text-align:right;"> 326.00 </td>
#>    <td style="text-align:right;"> 472.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> hp </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 146.688 </td>
#>    <td style="text-align:right;"> 68.563 </td>
#>    <td style="text-align:right;"> 52.000 </td>
#>    <td style="text-align:right;"> 96.500 </td>
#>    <td style="text-align:right;"> 123.000 </td>
#>    <td style="text-align:right;"> 180.00 </td>
#>    <td style="text-align:right;"> 335.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> drat </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 3.597 </td>
#>    <td style="text-align:right;"> 0.535 </td>
#>    <td style="text-align:right;"> 2.760 </td>
#>    <td style="text-align:right;"> 3.080 </td>
#>    <td style="text-align:right;"> 3.695 </td>
#>    <td style="text-align:right;"> 3.92 </td>
#>    <td style="text-align:right;"> 4.930 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> wt </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 3.217 </td>
#>    <td style="text-align:right;"> 0.978 </td>
#>    <td style="text-align:right;"> 1.513 </td>
#>    <td style="text-align:right;"> 2.581 </td>
#>    <td style="text-align:right;"> 3.325 </td>
#>    <td style="text-align:right;"> 3.61 </td>
#>    <td style="text-align:right;"> 5.424 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> qsec </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 17.849 </td>
#>    <td style="text-align:right;"> 1.787 </td>
#>    <td style="text-align:right;"> 14.500 </td>
#>    <td style="text-align:right;"> 16.892 </td>
#>    <td style="text-align:right;"> 17.710 </td>
#>    <td style="text-align:right;"> 18.90 </td>
#>    <td style="text-align:right;"> 22.900 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> vs </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 0.438 </td>
#>    <td style="text-align:right;"> 0.504 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 1.00 </td>
#>    <td style="text-align:right;"> 1.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> am </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 0.406 </td>
#>    <td style="text-align:right;"> 0.499 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 0.000 </td>
#>    <td style="text-align:right;"> 1.00 </td>
#>    <td style="text-align:right;"> 1.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> gear </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 3.688 </td>
#>    <td style="text-align:right;"> 0.738 </td>
#>    <td style="text-align:right;"> 3.000 </td>
#>    <td style="text-align:right;"> 3.000 </td>
#>    <td style="text-align:right;"> 4.000 </td>
#>    <td style="text-align:right;"> 4.00 </td>
#>    <td style="text-align:right;"> 5.000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> carb </td>
#>    <td style="text-align:right;"> 32 </td>
#>    <td style="text-align:right;"> 2.812 </td>
#>    <td style="text-align:right;"> 1.615 </td>
#>    <td style="text-align:right;"> 1.000 </td>
#>    <td style="text-align:right;"> 2.000 </td>
#>    <td style="text-align:right;"> 2.000 </td>
#>    <td style="text-align:right;"> 4.00 </td>
#>    <td style="text-align:right;"> 8.000 </td>
#>   </tr>
#> </tbody>
#> </table>