Prepares a Table Displaying Extreme Observations

Reads a data frame containing the variable of interest as the last variable, sorts it by it and displays the top and bottom n variables.

prepare_ext_obs_table(df, n = 5, ...)

Arguments

df
Data frame containing at least one variable that is numeric as variable of interest. Most likely it will also contain the variables identifying an observation.

n
The number of top/bottom observations that you want to report.

...
Additional parameters that are passed to kable.

Value

A list containing two items:

"df"

A data frame containing the top/bottom n observations

"kable_ret"

The return value provided by kable containing the formatted table

Details

The default parameters for calling kable, are format = "html", digits = 3, format.args = list(big.mark = ','), row.names = FALSE.

Examples

t <- prepare_ext_obs_table(data.frame(name = rownames(mtcars), hp = mtcars$hp))
t$df

#>                 name  hp
#> 31     Maserati Bora 335
#> 29    Ford Pantera L 264
#> 7         Duster 360 245
#> 24        Camaro Z28 245
#> 17 Chrysler Imperial 230
#> 18          Fiat 128  66
#> 26         Fiat X1-9  66
#> 20    Toyota Corolla  65
#> 8          Merc 240D  62
#> 19       Honda Civic  52