Prepares a Regression Table

Builds a regression table based on a set of user-specifed models or a single model and a partitioning variable.

prepare_regression_table(df, dvs, idvs, feffects = rep("", length(dvs)),
  clusters = rep("", length(dvs)), byvar = "", format = "html",
  drop_underscore = NULL)

Arguments

df
Data frame containing the data to estimate the models on.

dvs
A character vector containing the dependent variable(s).

idvs
A character vector or a a list of character vectors containing the independent variables.

feffects
A character vector or a a list of character vectors containing the fixed effects.

clusters
A character vector or a a list of character vectors containing the cluster variables.

byvar
A factorial variable to estimate the model on (only possible if only one model is being estimated).

format
A character scalar that is passed on stargazer as type to determine the presentation format (e.g., "html", "text", or "latex").

drop_underscore
A quick'n'dirty hack to address a bug in stargazer that triggers it to choke on underscores in variable names. If not NULL, all underscores in variable names will be replaced by the given string,

Value

A list contining two items

"models"

A list containg the model results and byvalues if appropriate

"table"

The output of stargazer containing the table

Details

This is a wrapper function calling the stargazer package. Depending on whether the dependent variable is numeric or a factor with two levels, the models are estimated using felm or glm (with family = binomial(link="logit")). Fixed effects and clustered standard errors are only supported with continous dependent variables. If run with byvar, only levels that have more observations than coefficients are estimated.

Examples

df <- data.frame(year = as.factor(floor(stats::time(datasets::EuStockMarkets))),
                 datasets::EuStockMarkets)
dvs = c("DAX", "SMI", "CAC", "FTSE")
idvs = list(c("SMI", "CAC", "FTSE"),
            c("DAX", "CAC", "FTSE"),
            c("SMI", "DAX", "FTSE"),
            c("SMI", "CAC", "DAX"))
feffects = list("year", "year", "year", "year")
clusters = list("year", "year", "year", "year")
t <- prepare_regression_table(df, dvs, idvs, feffects, clusters)
t$table

#>  [1] ""                                                                                                                                                                                                                                                 
#>  [2] "<table style=\"text-align:center\"><tr><td colspan=\"5\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\"></td><td colspan=\"4\"><em>Dependent variable:</em></td></tr>"                                      
#>  [3] "<tr><td></td><td colspan=\"4\" style=\"border-bottom: 1px solid black\"></td></tr>"                                                                                                                                                               
#>  [4] "<tr><td style=\"text-align:left\"></td><td>DAX</td><td>SMI</td><td>CAC</td><td>FTSE</td></tr>"                                                                                                                                                    
#>  [5] "<tr><td style=\"text-align:left\"></td><td>(1)</td><td>(2)</td><td>(3)</td><td>(4)</td></tr>"                                                                                                                                                     
#>  [6] "<tr><td colspan=\"5\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\">SMI</td><td>0.447<sup>***</sup></td><td></td><td>-0.095</td><td>0.566<sup>***</sup></td></tr>"                                         
#>  [7] "<tr><td style=\"text-align:left\"></td><td>(0.046)</td><td></td><td>(0.088)</td><td>(0.107)</td></tr>"                                                                                                                                            
#>  [8] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                 
#>  [9] "<tr><td style=\"text-align:left\">DAX</td><td></td><td>0.864<sup>***</sup></td><td>0.664<sup>***</sup></td><td>-0.188</td></tr>"                                                                                                                  
#> [10] "<tr><td style=\"text-align:left\"></td><td></td><td>(0.169)</td><td>(0.141)</td><td>(0.181)</td></tr>"                                                                                                                                            
#> [11] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                 
#> [12] "<tr><td style=\"text-align:left\">CAC</td><td>0.735<sup>***</sup></td><td>-0.204</td><td></td><td>0.188<sup>*</sup></td></tr>"                                                                                                                    
#> [13] "<tr><td style=\"text-align:left\"></td><td>(0.111)</td><td>(0.219)</td><td></td><td>(0.102)</td></tr>"                                                                                                                                            
#> [14] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                 
#> [15] "<tr><td style=\"text-align:left\">FTSE</td><td>-0.124</td><td>0.722<sup>***</sup></td><td>0.112</td><td></td></tr>"                                                                                                                               
#> [16] "<tr><td style=\"text-align:left\"></td><td>(0.125)</td><td>(0.060)</td><td>(0.075)</td><td></td></tr>"                                                                                                                                            
#> [17] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                 
#> [18] "<tr><td colspan=\"5\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\">Fixed effects</td><td>year</td><td>year</td><td>year</td><td>year</td></tr>"                                                           
#> [19] "<tr><td style=\"text-align:left\">Std. errors clustered</td><td>year</td><td>year</td><td>year</td><td>year</td></tr>"                                                                                                                            
#> [20] "<tr><td style=\"text-align:left\">Observations</td><td>1,860</td><td>1,860</td><td>1,860</td><td>1,860</td></tr>"                                                                                                                                 
#> [21] "<tr><td style=\"text-align:left\">R<sup>2</sup></td><td>0.994</td><td>0.995</td><td>0.981</td><td>0.989</td></tr>"                                                                                                                                
#> [22] "<tr><td style=\"text-align:left\">Adjusted R<sup>2</sup></td><td>0.994</td><td>0.995</td><td>0.981</td><td>0.989</td></tr>"                                                                                                                       
#> [23] "<tr><td colspan=\"5\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\"><em>Note:</em></td><td colspan=\"4\" style=\"text-align:right\"><sup>*</sup>p<0.1; <sup>**</sup>p<0.05; <sup>***</sup>p<0.01</td></tr>"
#> [24] "</table>"                                                                                                                                                                                                                                         

t <- prepare_regression_table(df, "DAX", c("SMI", "CAC", "FTSE"), byvar="year")
print(t$table)

#>  [1] ""                                                                                                                                                                                                                                                                                                                                                                                       
#>  [2] "<table style=\"text-align:center\"><tr><td colspan=\"10\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\"></td><td colspan=\"9\"><em>Dependent variable:</em></td></tr>"                                                                                                                                                                           
#>  [3] "<tr><td></td><td colspan=\"9\" style=\"border-bottom: 1px solid black\"></td></tr>"                                                                                                                                                                                                                                                                                                     
#>  [4] "<tr><td style=\"text-align:left\"></td><td colspan=\"9\">DAX</td></tr>"                                                                                                                                                                                                                                                                                                                 
#>  [5] "<tr><td style=\"text-align:left\"></td><td>Full Sample</td><td>1991</td><td>1992</td><td>1993</td><td>1994</td><td>1995</td><td>1996</td><td>1997</td><td>1998</td></tr>"                                                                                                                                                                                                               
#>  [6] "<tr><td style=\"text-align:left\"></td><td>(1)</td><td>(2)</td><td>(3)</td><td>(4)</td><td>(5)</td><td>(6)</td><td>(7)</td><td>(8)</td><td>(9)</td></tr>"                                                                                                                                                                                                                               
#>  [7] "<tr><td colspan=\"10\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\">SMI</td><td>0.493<sup>***</sup></td><td>0.634<sup>***</sup></td><td>-0.283<sup>***</sup></td><td>0.445<sup>***</sup></td><td>-0.265<sup>***</sup></td><td>0.333<sup>***</sup></td><td>0.189<sup>***</sup></td><td>0.258<sup>***</sup></td><td>0.244<sup>***</sup></td></tr>"
#>  [8] "<tr><td style=\"text-align:left\"></td><td>(0.015)</td><td>(0.068)</td><td>(0.056)</td><td>(0.021)</td><td>(0.039)</td><td>(0.043)</td><td>(0.020)</td><td>(0.026)</td><td>(0.043)</td></tr>"                                                                                                                                                                                           
#>  [9] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                                                                                                          
#> [10] "<tr><td style=\"text-align:left\">CAC</td><td>0.496<sup>***</sup></td><td>-0.052</td><td>0.952<sup>***</sup></td><td>0.516<sup>***</sup></td><td>0.721<sup>***</sup></td><td>0.304<sup>***</sup></td><td>0.577<sup>***</sup></td><td>1.019<sup>***</sup></td><td>1.182<sup>***</sup></td></tr>"                                                                                         
#> [11] "<tr><td style=\"text-align:left\"></td><td>(0.015)</td><td>(0.045)</td><td>(0.026)</td><td>(0.033)</td><td>(0.052)</td><td>(0.060)</td><td>(0.039)</td><td>(0.075)</td><td>(0.054)</td></tr>"                                                                                                                                                                                           
#> [12] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                                                                                                          
#> [13] "<tr><td style=\"text-align:left\">FTSE</td><td>-0.017</td><td>0.066</td><td>-0.069<sup>**</sup></td><td>0.134<sup>***</sup></td><td>-0.085<sup>**</sup></td><td>0.017</td><td>0.389<sup>***</sup></td><td>0.195<sup>***</sup></td><td>-0.378<sup>***</sup></td></tr>"                                                                                                                   
#> [14] "<tr><td style=\"text-align:left\"></td><td>(0.021)</td><td>(0.046)</td><td>(0.035)</td><td>(0.040)</td><td>(0.039)</td><td>(0.049)</td><td>(0.027)</td><td>(0.045)</td><td>(0.045)</td></tr>"                                                                                                                                                                                           
#> [15] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                                                                                                          
#> [16] "<tr><td style=\"text-align:left\">Constant</td><td>-175.946<sup>***</sup></td><td>458.772<sup>***</sup></td><td>575.676<sup>***</sup></td><td>-688.784<sup>***</sup></td><td>1,623.337<sup>***</sup></td><td>567.057<sup>***</sup></td><td>-803.680<sup>***</sup></td><td>-1,353.434<sup>***</sup></td><td>1,146.634<sup>***</sup></td></tr>"                                           
#> [17] "<tr><td style=\"text-align:left\"></td><td>(44.666)</td><td>(72.607)</td><td>(72.806)</td><td>(56.941)</td><td>(65.786)</td><td>(105.476)</td><td>(68.069)</td><td>(110.585)</td><td>(174.146)</td></tr>"                                                                                                                                                                               
#> [18] "<tr><td style=\"text-align:left\"></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>"                                                                                                                                                                                                                                                          
#> [19] "<tr><td colspan=\"10\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\">Fixed effects</td><td>None</td><td>None</td><td>None</td><td>None</td><td>None</td><td>None</td><td>None</td><td>None</td><td>None</td></tr>"                                                                                                                               
#> [20] "<tr><td style=\"text-align:left\">Std. errors clustered</td><td>No</td><td>No</td><td>No</td><td>No</td><td>No</td><td>No</td><td>No</td><td>No</td><td>No</td></tr>"                                                                                                                                                                                                                   
#> [21] "<tr><td style=\"text-align:left\">Observations</td><td>1,860</td><td>131</td><td>260</td><td>260</td><td>260</td><td>260</td><td>260</td><td>260</td><td>169</td></tr>"                                                                                                                                                                                                                 
#> [22] "<tr><td style=\"text-align:left\">R<sup>2</sup></td><td>0.990</td><td>0.669</td><td>0.850</td><td>0.973</td><td>0.560</td><td>0.720</td><td>0.932</td><td>0.964</td><td>0.978</td></tr>"                                                                                                                                                                                                
#> [23] "<tr><td style=\"text-align:left\">Adjusted R<sup>2</sup></td><td>0.990</td><td>0.661</td><td>0.848</td><td>0.973</td><td>0.555</td><td>0.717</td><td>0.931</td><td>0.964</td><td>0.978</td></tr>"                                                                                                                                                                                       
#> [24] "<tr><td colspan=\"10\" style=\"border-bottom: 1px solid black\"></td></tr><tr><td style=\"text-align:left\"><em>Note:</em></td><td colspan=\"9\" style=\"text-align:right\"><sup>*</sup>p<0.1; <sup>**</sup>p<0.05; <sup>***</sup>p<0.01</td></tr>"                                                                                                                                     
#> [25] "</table>"