Prepares a Regression Table

Builds a regression table based on a set of user-specified models or a single model and a partitioning variable.

prepare_regression_table(
  df,
  dvs,
  idvs,
  feffects = rep("", length(dvs)),
  clusters = rep("", length(dvs)),
  models = rep("auto", length(dvs)),
  byvar = "",
  format = "html"
)

Arguments

df
Data frame containing the data to estimate the models on.

dvs
A character vector containing the variable names for the dependent variable(s).

idvs
A character vector or a a list of character vectors containing the variable names of the independent variables.

feffects
A character vector or a a list of character vectors containing the variable names of the fixed effects.

clusters
A character vector or a a list of character vectors containing the variable names of the cluster variables.

models
A character vector indicating the model types to be estimated ('ols', 'logit', or 'auto')

byvar
A factorial variable to estimate the model on (only possible if only one model is being estimated).

format
A character scalar that is passed on stargazer as type to determine the presentation format ("html", "text", or "latex").

Value

A list containing two items

"models"

A list containing the model results and by values if appropriate

"table"

The output of stargazer containing the table

Details

This is a wrapper function calling the stargazer package. For numeric dependent variables the models are estimated using lm for models without and plm for models with fixed effects. Binary dependent variable models are estimated using glm (with family = binomial(link="logit")). You can override this behavior by specifying the model with the models parameter. Multinomial logit models are not supported. Clustered standard errors are estimated using plm's robust covariance matrix estimators for OLS and cluster.vcov for logit models. Only up to two dimensions are supported for fixed effects and standard error clusters need to be also present as fixed effects. If run with byvar, only levels that have more observations than coefficients are estimated.

Examples

df <- data.frame(year = as.factor(floor(stats::time(datasets::EuStockMarkets))),
                 datasets::EuStockMarkets)
dvs = c("DAX", "SMI", "CAC", "FTSE")
idvs = list(c("SMI", "CAC", "FTSE"),
            c("DAX", "CAC", "FTSE"),
            c("SMI", "DAX", "FTSE"),
            c("SMI", "CAC", "DAX"))
feffects = list("year", "year", "year", "year")
clusters = list("year", "year", "year", "year")
t <- prepare_regression_table(df, dvs, idvs, feffects, clusters, format = "text")
t$table

#>  [1] ""                                                         
#>  [2] "========================================================="
#>  [3] "                              Dependent variable:        "
#>  [4] "                      -----------------------------------"
#>  [5] "                        DAX      SMI      CAC      FTSE  "
#>  [6] "                        (1)      (2)      (3)      (4)   "
#>  [7] "---------------------------------------------------------"
#>  [8] "SMI                   0.447***           -0.095  0.566***"
#>  [9] "                      (0.046)           (0.088)  (0.107) "
#> [10] "                                                         "
#> [11] "DAX                            0.864*** 0.664***  -0.188 "
#> [12] "                               (0.169)  (0.141)  (0.180) "
#> [13] "                                                         "
#> [14] "CAC                   0.735***  -0.204            0.188* "
#> [15] "                      (0.111)  (0.218)           (0.102) "
#> [16] "                                                         "
#> [17] "FTSE                   -0.124  0.722***  0.112           "
#> [18] "                      (0.125)  (0.060)  (0.075)          "
#> [19] "                                                         "
#> [20] "---------------------------------------------------------"
#> [21] "Estimator               ols      ols      ols      ols   "
#> [22] "Fixed effects           year     year     year     year  "
#> [23] "Std. errors clustered   year     year     year     year  "
#> [24] "Observations           1,860    1,860    1,860    1,860  "
#> [25] "R2                     0.900    0.885    0.811    0.747  "
#> [26] "Adjusted R2            0.900    0.884    0.810    0.746  "
#> [27] "========================================================="
#> [28] "Note:                         *p<0.1; **p<0.05; ***p<0.01"

t <- prepare_regression_table(df, "DAX", c("SMI", "CAC", "FTSE"), byvar="year", format = "text")
print(t$table)

#>  [1] ""                                   "% Error: Unrecognized object type."
#>  [3] "% Error: Unrecognized object type." "% Error: Unrecognized object type."
#>  [5] "% Error: Unrecognized object type." "% Error: Unrecognized object type."
#>  [7] "% Error: Unrecognized object type." "% Error: Unrecognized object type."
#>  [9] "% Error: Unrecognized object type." "% Error: Unrecognized object type."